requirements_dev.txt
pytest-asyncio==1.3.0
pytest==8.3.4
pytest-aiohttp==1.1.0
